requirements.txt
asgiref==3.8.1
Django==5.1.5
django-modeltranslation==0.19.12
pillow==11.1.0
psycopg2==2.9.10
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
